extract less vars
